Code cleanup and enum performance enhancements

+++ src/AngleSharp/Css/Values/Angle.cs
@@ -256,7 +256,7 @@
         /// <summary>
         /// An enumeration of angle representations.
         /// </summary>
-        public enum Unit : ushort
+        public enum Unit : byte
         {
             /// <summary>
             /// No valid unit.
